requirements.txt
numpy
opencv-python-headless
Pillow
protobuf
python-dotenv
streamlit
google-generativeai
